LayoutTests/platform/chromium-win/platform/chromium/fast/forms/time/time-suggestion-picker-key-operations-expected.txt

Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
Check that page popup exists.
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
Check that up/down arrow keys work.
PASS highlightedEntry() is null
PASS highlightedEntry() is "01:01"
PASS highlightedEntry() is "01:02"
PASS highlightedEntry() is "01:01"
Check that up arrow key at top of list does nothing.
PASS highlightedEntry() is "01:01"
PASS highlightedEntry() is "01:02"
Check that page up/down arrow keys work.
PASS highlightedEntry() is "01:40"
PASS highlightedEntry() is "02:00"
PASS highlightedEntry() is "01:21"
PASS highlightedEntry() is "02:00"
PASS highlightedEntry() is "02:08"
Check that down arrow key at bottom of list does nothing.
PASS highlightedEntry() is "02:08"
Check that enter key sets the highlighted value.
PASS highlightedEntry() is "02:07"
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("time").value is "02:07"
Reopen popup.
Check that escape key closes the popup.
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("time").value is "02:07"
PASS F4 opened picker.
PASS successfullyParsed is true

TEST COMPLETE
 
